log4cpp: renamed library and pkg-config file to orocos-log4cpp and set version to 2.9.0
This patch allows to install the patched version of log4cpp alongside the original log4cpp
package without potential conflicts and without the need to increase the version number to 6.0.0
as a criteria to detect the patched log4cpp library in ocl.

rtt and ocl needs a minimal patch to change the library name and to remove the version 6.0.0 check.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>